+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Detect Epic Games installation folder
 - Add transition when changing vehicle modes
+- Add "landing gear deployed" ambient mode
 
 ### Changed